pxlcolor - emit zero sized clip rectangles
Bug #707687 "pxlcolor device fails to clip content"

Commit 4df181c0f8323b84239cb46535494835931994b4 adds clipping to the
xpswrite device, and also dropped empty (singlusar) clip rectangles.

It seems that doing this causes the vector device to draw images that
should be clipped out, because the emoty clip is not written to the
file. This only seems to cause a problem with the pxlcolor device,
possibly because other high level devices don't rely on the vector
device to do this work.

The simple solution is to write out the empty clip paths as we did
before.